Refactor "phones" to "suite" to reflect project changes
The start of a new world, changing our focus from just finding phone
numbers, to handling data overall
