+++ bin/.gitignore
@@ -1,1 +1,2 @@
 /phones/
+/suite/
